ADD utils/network and extracted FetchFromHTTPResource. ADD types.go and extracted types from yaml. UPDATE internal/yaml getBigBangValuesYamlByes renamed to getBigBangValuesYaml after adding unmarshalling to this method and changing the return type to BigBangValues. Removed getBigBangValues and replaced usage with getBigBangValuesYaml.
